+++ Gemfile
@@ -17,6 +17,7 @@ gem "codeclimate-test-reporter", group: :test, require: nil
 gem 'bcrypt', '~> 3.1.7'
 gem 'active_model_serializers', '~> 0.10.0'
 gem 'figaro'
+gem 'materialize-sass', '~> 0.97.7'
 
 group :development, :test do
   gem 'byebug'
