+++ config/routes.rb
@@ -2,6 +2,14 @@ require 'api_constraints'
 
 Rails.application.routes.draw do
 
+  resources :users, only: [:create, :show, :update, :destroy]
+
+  scope controller: :users do
+    get '/signup' => :new, as: :signup
+    post 'create' => :create
+    get '/show' => :show
+  end
+
   scope module: :api, defaults: { format: :json } do
     scope module: :v1 do
       constraints ApiConstraints.new(version: 1, default: true) do
@@ -9,7 +17,6 @@ Rails.application.routes.draw do
           post 'auth/login' => :login
           get 'auth/logout' => :logout
         end
-        resources :users, only: [:create, :show, :update, :destroy]
         resources :bucketlists do
           resources :items
         end
