+++ spec/controllers/users_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Api::V1::UsersController, type: :controller do
+RSpec.describe UsersController, type: :controller do
   describe 'GET #show' do
     let(:user) { FactoryGirl.create :user }
     before { get :show, id: user.id }
